Refactor the logic to get plugin info into a new hook
