+++ client/my-sites/plugins/plugin-management-v2/update-plugin/index.tsx
@@ -2,12 +2,9 @@ import { Button } from '@automattic/components';
 import { Icon, arrowRight } from '@wordpress/icons';
 import classNames from 'classnames';
 import { useTranslate } from 'i18n-calypso';
-import { useMemo } from 'react';
 import { UPDATE_PLUGIN } from 'calypso/lib/plugins/constants';
-import { siteObjectsToSiteIds } from 'calypso/my-sites/plugins/utils';
 import { useSelector } from 'calypso/state';
-import { getPluginOnSites } from 'calypso/state/plugins/installed/selectors';
-import getSites from 'calypso/state/selectors/get-sites';
+import usePluginVersionInfo from '../hooks/use-plugin-version-info';
 import PluginActionStatus from '../plugin-action-status';
 import { getAllowedPluginActions } from '../utils/get-allowed-plugin-actions';
 import { getPluginActionStatuses } from '../utils/get-plugin-action-statuses';
@@ -25,66 +22,12 @@ interface Props {
 
 export default function UpdatePlugin( { plugin, selectedSite, className, updatePlugin }: Props ) {
 	const translate = useTranslate();
-	const allSites = useSelector( getSites );
 	const state = useSelector( ( state ) => state );
 
-	const getPluginSites = ( plugin: PluginComponentProps ) => {
-		return Object.keys( plugin.sites ).map( ( siteId ) => {
-			const site = allSites.find( ( s ) => s?.ID === parseInt( siteId ) );
-			return {
-				...site,
-				...plugin.sites[ siteId ],
-			} as any; // This must be cast as any until this file is updated to work with the selectors in state/plugins/installed/selectors
-		} );
-	};
-
-	const sites = getPluginSites( plugin );
-	const siteIds = siteObjectsToSiteIds( sites );
-	const pluginsOnSites: any = getPluginOnSites( state, siteIds, plugin?.slug );
-
-	const currentVersions = sites
-		.map( ( site ) => {
-			const siteId = selectedSite ? selectedSite.ID : site.ID;
-			const sitePlugin = pluginsOnSites?.sites[ siteId ];
-			return sitePlugin?.version;
-		} )
-		.filter( ( version ) => version );
-
-	const updatedVersions = sites
-		.map( ( site ) => {
-			const siteId = selectedSite ? selectedSite.ID : site.ID;
-			const sitePlugin = pluginsOnSites?.sites[ siteId ];
-			return sitePlugin?.update?.new_version;
-		} )
-		.filter( ( version ) => version );
-
-	const currentVersionsRange = useMemo( () => {
-		const versions = [
-			// We want to remove the duplicated versions in the array, because if multiple sites have
-			// the same plugin version, we don't want to display the range.
-			...new Set(
-				// Sort the plugin versions, respecting semantic version convention.
-				currentVersions.sort( ( a: string, b: string ): number =>
-					a.localeCompare( b, undefined, {
-						numeric: true,
-						sensitivity: 'case',
-						caseFirst: 'upper',
-					} )
-				)
-			),
-		];
-
-		return {
-			min: versions[ 0 ],
-			max: versions.length > 1 ? versions[ versions.length - 1 ] : null,
-		};
-	}, [ currentVersions ] );
-
-	const hasUpdate = sites.some( ( site ) => {
-		const siteId = selectedSite ? selectedSite.ID : site.ID;
-		const sitePlugin = pluginsOnSites?.sites[ siteId ];
-		return sitePlugin?.update?.new_version && site.canUpdateFiles;
-	} );
+	const { currentVersionsRange, updatedVersions, hasUpdate } = usePluginVersionInfo(
+		plugin,
+		selectedSite
+	);
 
 	const allowedActions = getAllowedPluginActions( plugin, state, selectedSite );
